renamed pageStore to UpdateUIstore
